Large refactoring work in preparation for adding concurrent animations.
