--- DrawerKit/DrawerKit/Internal API/DrawerState.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-enum DrawerState: Equatable {
-    case collapsed
-    case partiallyExpanded
-    case fullyExpanded
-    case transitioning(CGFloat) // current drawer Y position
-}
