+++ DrawerKit/DrawerKit/Internal API/PresentationController+Gestures.swift
@@ -1,0 +1,44 @@
+import UIKit
+
+extension PresentationController {
+    @objc func handleDrawerFullExpansionTap() {
+        guard let tapGesture = drawerFullExpansionTapGR else { return }
+        let tapY = tapGesture.location(in: presentedView).y
+        guard tapY < drawerPartialH else { return }
+        animateTransition(to: .fullyExpanded)
+    }
+
+    @objc func handleDrawerDismissalTap() {
+        guard let tapGesture = drawerDismissalTapGR else { return }
+        let tapY = tapGesture.location(in: containerView).y
+        guard tapY < currentDrawerY else { return }
+        presentedViewController.dismiss(animated: true)
+    }
+
+    @objc func handleDrawerDrag() {
+        guard let panGesture = drawerDragGR, let view = panGesture.view else { return }
+
+        switch panGesture.state {
+        case .began:
+            lastDrawerState = drawerState(for: currentDrawerY, clampToNearest: true)
+
+        case .changed:
+            lastDrawerState = drawerState(for: currentDrawerY, clampToNearest: true)
+            currentDrawerY += panGesture.translation(in: view).y
+            currentDrawerCornerRadius = cornerRadius(at: currentDrawerState)
+            panGesture.setTranslation(.zero, in: view)
+
+        case .ended:
+            let drawerSpeedY = panGesture.velocity(in: view).y / containerViewH
+            let endingState = nextStateFrom(currentState: currentDrawerState,
+                                            speedY: drawerSpeedY)
+            animateTransition(to: endingState)
+
+        case .cancelled:
+            animateTransition(to: lastDrawerState)
+
+        default:
+            break
+        }
+    }
+}
