+++ DrawerKit/DrawerKit/Internal API/PresentationController+Setup.swift
@@ -1,0 +1,88 @@
+import UIKit
+
+extension PresentationController {
+    func setupDrawerFullExpansionTapRecogniser() {
+        guard drawerFullExpansionTapGR == nil else { return }
+        let isFullyPresentable = isFullyPresentableByDrawerTaps
+        let numTapsRequired = numberOfTapsForFullDrawerPresentation
+        guard isFullyPresentable && numTapsRequired > 0 else { return }
+        let tapGesture = UITapGestureRecognizer(target: self,
+                                                action: #selector(handleDrawerFullExpansionTap))
+        tapGesture.numberOfTouchesRequired = 1
+        tapGesture.numberOfTapsRequired = numTapsRequired
+        presentedView?.addGestureRecognizer(tapGesture)
+        drawerFullExpansionTapGR = tapGesture
+    }
+
+    func removeDrawerFullExpansionTapRecogniser() {
+        guard let tapGesture = drawerFullExpansionTapGR else { return }
+        presentedView?.removeGestureRecognizer(tapGesture)
+        drawerFullExpansionTapGR = nil
+    }
+}
+
+extension PresentationController {
+    func setupDrawerDismissalTapRecogniser() {
+        guard drawerDismissalTapGR == nil else { return }
+        let isDismissable = isDismissableByOutsideDrawerTaps
+        let numTapsRequired = numberOfTapsForOutsideDrawerDismissal
+        guard isDismissable && numTapsRequired > 0 else { return }
+        let tapGesture = UITapGestureRecognizer(target: self,
+                                                action: #selector(handleDrawerDismissalTap))
+        tapGesture.numberOfTouchesRequired = 1
+        tapGesture.numberOfTapsRequired = numTapsRequired
+        containerView?.addGestureRecognizer(tapGesture)
+        drawerDismissalTapGR = tapGesture
+    }
+
+    func removeDrawerDismissalTapRecogniser() {
+        guard let tapGesture = drawerDismissalTapGR else { return }
+        containerView?.removeGestureRecognizer(tapGesture)
+        drawerDismissalTapGR = nil
+    }
+}
+
+extension PresentationController {
+    func setupDrawerDragRecogniser() {
+        guard drawerDragGR == nil && isDrawerDraggable else { return }
+        let panGesture = UIPanGestureRecognizer(target: self,
+                                                action: #selector(handleDrawerDrag))
+        presentedView?.addGestureRecognizer(panGesture)
+        drawerDragGR = panGesture
+    }
+
+    func removeDrawerDragRecogniser() {
+        guard let panGesture = drawerDragGR else { return }
+        presentedView?.removeGestureRecognizer(panGesture)
+        drawerDragGR = nil
+    }
+}
+
+extension PresentationController {
+    func setupDebugHeightMarks() {
+        guard inDebugMode && (upperMarkGap > 0 || lowerMarkGap > 0),
+            let containerView = containerView else { return }
+
+        if upperMarkGap > 0 {
+            let upperMarkYView = UIView()
+            upperMarkYView.backgroundColor = .black
+            upperMarkYView.frame = CGRect(x: 0, y: upperMarkY,
+                                          width: containerView.bounds.size.width, height: 3)
+            containerView.addSubview(upperMarkYView)
+        }
+
+        if lowerMarkGap > 0 {
+            let lowerMarkYView = UIView()
+            lowerMarkYView.backgroundColor = .black
+            lowerMarkYView.frame = CGRect(x: 0, y: lowerMarkY,
+                                          width: containerView.bounds.size.width, height: 3)
+            containerView.addSubview(lowerMarkYView)
+        }
+
+        let drawerMarkView = UIView()
+        drawerMarkView.backgroundColor = .white
+        drawerMarkView.frame = CGRect(x: 0, y: drawerPartialY,
+                                      width: containerView.bounds.size.width, height: 3)
+        containerView.addSubview(drawerMarkView)
+    }
+}
