+++ DrawerKit/DrawerKit/Public API/Protocols/DrawerPresentable.swift
@@ -1,6 +1,7 @@
 import UIKit
 
-/// Protocol that view controllers presented inside a drawer must conform to.
+/// A protocol that view controllers presented inside a drawer must conform to.
+
 public protocol DrawerPresentable: class {
     /// The height at which the drawer must be presented when it's in its
     /// partially expanded state. If negative, its value is clamped to zero.
