+++ DrawerKitDemo/DrawerKitDemo/PresenterViewController.swift
@@ -19,8 +19,11 @@ private extension PresenterViewController {
 
         // ... or after initialisation. All of these have default values so change only
         // what you need to configure differently. They're all listed here just so you
-        // can see what can be configured.
-        configuration.totalDurationInSeconds = 0.3
+        // can see what can be configured. The values listed are the default ones,
+        // except where indicated otherwise.
+        configuration.totalDurationInSeconds = 0.4
+        configuration.durationIsProportionalToDistanceTraveled = false
+        // default is UISpringTimingParameters()
         configuration.timingCurveProvider = UISpringTimingParameters(dampingRatio: 0.8)
         configuration.supportsPartialExpansion = true
         configuration.dismissesInStages = true
@@ -30,9 +33,9 @@ private extension PresenterViewController {
         configuration.isDismissableByOutsideDrawerTaps = true
         configuration.numberOfTapsForOutsideDrawerDismissal = 1
         configuration.flickSpeedThreshold = 3
-        configuration.upperMarkGap = 100
-        configuration.lowerMarkGap = 80
-        configuration.maximumCornerRadius = 20
+        configuration.upperMarkGap = 100 // default is 40
+        configuration.lowerMarkGap =  80 // default is 40
+        configuration.maximumCornerRadius = 15
 
         drawerDisplayController = DrawerDisplayController(presentingViewController: self,
                                                           presentedViewController: vc,
